+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# Upcoming
+
+- Fixed tab bar stencil and damage rendering
+
 # hl0.53.0 and before
 
 - Fixed tab bar not being damaged (glitching) on window title change.